backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
opencv-python==4.10.0.84
ultralytics==8.3.0
# PyTorch - Will auto-detect platform (Apple Silicon MPS, CUDA, or CPU)
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4
pillow==10.4.0
